kernel/sched/fair.c: Fix divide by zero
This issue has been reported (not so frequently), but occasionally. The
real issue is divisor overflow, subtle overflow. We never expect
delta to be greater than sched_avg_period(), but there is a catch -
this can happen if nohz is turned on, effectively that means we don't
update age_stamp

        delta = __rq_clock_broken(rq) - age_stamp;

        if (unlikely(delta < 0))
                delta = 0;

        total = sched_avg_period() + delta;

age_stamp is updated in sched_update_avg() which could be skipped
if nohz is enabled. My calculations show that a delta value of
0xfe2329b00 in delta will cause a divide by zero.

Signed-off-by: Balbir Singh <sblbir@amazon.com>
Reviewed-by: Andy Strohman <astroh@amazon.com>
Reviewed-by: Samuel Mendoza-Jonas <samjonas@amazon.com>